+++ tests/src/main/java/com/github/klee0kai/test/di/techfactory/TechFactoryComponent.java
@@ -1,4 +1,4 @@
-package com.github.klee0kai.test.di.phone_factory;
+package com.github.klee0kai.test.di.techfactory;
 
 import com.github.klee0kai.stone.annotations.component.Component;
